Do proper clipping when dumping instead of just bounding box checks.
(The actual coordinates of the polyline won't be clipped.  The clipping
is just to check whether it should be dumped with this tile at all.)

It doesn't actually make a huge difference, but it does exclude some
gigantic vectors at low-numbered zoom levels that have coordinates on
both sides of the tile in question but don't actually intersect it.